@@ -1,21 +1,21 @@
 module tests
 
-go 1.23
+go 1.24
 
-toolchain go1.23.2
+toolchain go1.24.0
 
 require (
-	github.com/roadrunner-server/config/v5 v5.0.4
-	github.com/roadrunner-server/endure/v2 v2.6.1
+	github.com/roadrunner-server/config/v5 v5.1.7
+	github.com/roadrunner-server/endure/v2 v2.6.2
 	github.com/roadrunner-server/errors v1.4.1
-	github.com/roadrunner-server/http/v5 v5.0.4
-	github.com/roadrunner-server/logger/v5 v5.0.4
-	github.com/roadrunner-server/metrics/v5 v5.0.4
-	github.com/roadrunner-server/pool v1.0.2
-	github.com/roadrunner-server/prometheus/v5 v5.0.3
-	github.com/roadrunner-server/rpc/v5 v5.0.4
+	github.com/roadrunner-server/http/v5 v5.2.6
+	github.com/roadrunner-server/logger/v5 v5.1.7
+	github.com/roadrunner-server/metrics/v5 v5.1.7
+	github.com/roadrunner-server/pool v1.1.3
+	github.com/roadrunner-server/prometheus/v5 v5.1.6
+	github.com/roadrunner-server/rpc/v5 v5.1.7
 	github.com/roadrunner-server/server/v5 v5.0.0
-	github.com/stretchr/testify v1.9.0
+	github.com/stretchr/testify v1.10.0
 	go.uber.org/zap v1.27.0
 )
 
