@@ -28,7 +28,7 @@ func newCommand(log *zap.Logger, appLog *zap.Logger, cfg *InitConfig) *command {
 
 func (b *command) start() error {
 	const op = errors.Op("server_on_init")
-	stopCh := make(chan struct{}, 1)
+	stopCh := make(chan error, 1)
 
 	cmd := b.createProcess(b.cfg.Env, b.cfg.Command)
 
@@ -50,9 +50,11 @@ func (b *command) start() error {
 		errW := cmd.Wait()
 		if errW != nil {
 			b.log.Error("process wait", zap.Error(errW))
+			stopCh <- errW
+			return
 		}
 
-		stopCh <- struct{}{}
+		stopCh <- nil
 	}()
 
 	select {
@@ -60,12 +62,18 @@ func (b *command) start() error {
 		err = cmd.Process.Kill()
 		if err != nil {
 			b.log.Error("process killed", zap.Error(err))
+			return err
 		}
+
+		if b.cfg.ExitOnError {
+			return errors.Str("startup process has been killed by timeout")
+		}
+
 		return nil
 
-	case <-stopCh:
+	case err := <-stopCh:
 		timer.Stop()
-		return nil
+		return err
 	}
 }
 
